"working" with other types of sensors but not light
This has a partial interface constructed -- I got as far as showing a
list of services. When I use
    const lightSensors = useServices({ serviceClass: SRV_REFLECTED_LIGHT})

lightSensors is empty. If I use something else, like SRV_ACCELEROMETER
or SRV_BUTTON, the result of useServices is not empty. Until I debug
that I cannot ensure that a light sensor is selected